requirements.txt
pdf2docx>=0.5.6
docx2pdf>=0.1.8       
Pillow>=10.0.0